+++ kibot/pre_pcb_replace.py
@@ -28,6 +28,7 @@ class PCB_ReplaceOptions(Base_ReplaceOptions):
 @pre_class
 class PCB_Replace(Base_Replace):  # noqa: F821
     """ [dict] Replaces tags in the schematic. I.e. to insert the git hash or last revision date.
+        This is useful for KiCad 5, use `set_text_variables` when using KiCad 6.
         This pre-flight modifies the PCB. Even when a back-up is done use it carefully """
     _context = 'PCB'